Move origin misplaced test cases to correct location. Merge original failing commit behavior test to correct behavior.
